refactor(BuySell): move quote input logic to sub component
